feat: allow ignoring empty results when saving to cache
This new feature allows users to futher customize the `DbQueryResultCache` provider by indicating
that new (not yet cached) queries should only be cached if they return a non-empty response.
This allows for the use case where users may want to search for a record and use it if it exists,
otherwise create it if it does not exist. Without this new feature, the cache would store the empty
response for the initial query for its duration, and the process would create a new record. However,
the creation does not update the cache, and therefore future calls of the query before the
expiration of the entry will falsy indicate the record does not exist and then the process logic
will attempt to create the record, which now already exists.